requirements.txt
boto3
infoblox-client
invoke
keystoneauth1
openstacksdk
python-glanceclient
python-keystoneclient
python-openstackclient
s3cmd
